MAINT: restructure package
