+++ irrad_control/processes/converter.py
@@ -9,7 +9,7 @@ from uncertainties import ufloat, unumpy
 # Package imports
 import irrad_control.analysis as analysis
 import irrad_control.devices.readout as ro
-from irrad_control.utils.daq_proc import DAQProcess
+from irrad_control.processes.daq import DAQProcess
 
 
 class IrradConverter(DAQProcess):
