+++ irrad_control/scripts/configure_server.sh
